refactor(log): add zap for DefaultLogger (#1)
To avoid excessive impact on the code base,
only the log part of the code is modified.
Code refactoring is realized by using encapsulation function.

However, the disadvantages are also obvious,
and the role of zap log is not really played.

Therefore, should consider adding `ZapLogger` instead of modifying `DefaultLogger.

1. add zap
2. benchmark it
3. use happy path for if